openpose_docker/requirements.txt
numpy==1.19.2
Flask==1.1.2
opencv_python==4.5.2.54
yacs==0.1.8
requests==2.22.0
Pillow==8.1.2
psutil==5.8.0
PyYAML==5.4.1
cython==0.29.21


